requirements.txt
colorama==0.4.6
docx2pdf==0.1.8
lxml==4.9.2
Pillow==9.5.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
pywin32==306
tqdm==4.65.0
XlsxWriter==1.0.5